requirements.txt
briefcase==0.3.15
paho-mqtt==1.6.1
PyQt5==5.15.9

